EM-7 Queue Object Queue, Timer and Tests
Progress on timer to execute periodically the process

Timer working and Executable Project to test

Executable Project reference

Basic operation of the Queue with a Timer

QueueItems are processed and moved to tail

The items are removed if their Status is OK

Test Console project to test the Queue

ThreadPool and added the Subscription

Also moved Subscription to Entities and replaced the one that was there with this one.

Removed first item

This change generated a speed boost because the Tick parses less items from the item List each time.
MoveToTail method removed because since the item never returns to the queue, is now unnecessary.

Adding ConcurrentQueue

This instead of a regular List that might cause thread errors during item access

Lock to boolean to start and stop Timer

Boolean is now an Int because Interlock uses ints

Added Debug

cleaning code

SubscriberConfig options

Each QueueItem has a SubscriberConfig with a MaxTries and a Rate per second, so we increment a Tries counter in the item and and we check if the time is longer than the last time it tried with the millisecondRate.

Update Queue.cs

Added Serilog